test(perf): migrate to new multi-store API
