+++ tests/perf/test-app/src/schema.ts
@@ -55,3 +55,5 @@ const materializers = State.SQLite.materializers(events, {
 const state = State.SQLite.makeState({ tables, materializers })
 
 export const schema = makeSchema({ events, state })
+
+export type AppSchema = typeof schema
